sources/requirements.txt
nptyping==1.3.0
Pillow==7.2.0
numpy==1.19.2
